Typo fix, Add missing native to include

+++ httpclient.cpp
@@ -144,7 +144,7 @@ void HTTPClient::SetTimeout(int timeout)
 
 int HTTPClient::GetMaxSendSpeed() const
 {
-	return this->maxSendSpeed;
+	return (int)this->maxSendSpeed;
 }
 
 void HTTPClient::SetMaxSendSpeed(int speed)
@@ -154,7 +154,7 @@ void HTTPClient::SetMaxSendSpeed(int speed)
 
 int HTTPClient::GetMaxRecvSpeed() const
 {
-	return this->maxRecvSpeed;
+	return (int)this->maxRecvSpeed;
 }
 
 void HTTPClient::SetMaxRecvSpeed(int speed)
